Initial commit of 3D SIM template
This is a template notebook for 3D SIM reconstruction.   The goal is to
create a notebook for which it is easy to adapt to new data and which is
repeatable from run to run and machine to machine.

This notebook can be modified for new data, new OTF or new machine by
changing the paths defined in the first 2 cells of the notebook.