Fix number c threads at 1 (for now)
The python tiled code starts a thread for each core, each thread calls
the c code, in turn the c code was starting a thread for each core.
This means it was possible to have num_cores*number_cores threads.  This
sometimes may have caused a speed issue if other processes were using
many cores.

Setting the number of c threads to 1, does not hurt the speed of the
tiled version.  It will hurt the speed of non-tiled a bit, so long term
the number of threads should be a variable.