Optionally find the second amplitude
If there are three orders amplitude for order 1 and order 2 are found
during optimization.  However previously only amplitude was output to
the text file.  That has been changed so that the second order is also
output now.  When processing text output we can optionaly search for
amp2 (it is optional because for older files it won't be there).