Add ampmin, ampmax, k0spacings as parameters
ampmin and ampmax are used to constrain the amplitudes.  This is useful
if we use very small tiles as in this case the amplutudes can be
unreliable.

It is useful to be able to explitly set all k0spacings for the
constrained tiled reconstruction.  Sometimes we want to re-run the tile
with all parameters fixed subject to constraints.

This commit also outputs the second order amplitudes to the text file.
This is also used for tiled reconstruction.  We need to check both
amplitudes because either can go out of range.